requirements.txt
flask==3.0.0
flask-login==0.6.3
psycopg2-binary==2.9.9
yfinance==0.2.18
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.4 